:diamonds: Extract repo names from rancher-compose.yml
The rancher-compose.yml file contains variables for each of the image
tags for the parent service and each dependent services.
Extract the dependent service names from from this file.

It does rely on the naming convention for the variables to be
maintained in that the format
i.e. {sanitisied repo name}_docker_image_tag
sanitisation takes the form of translating '-' to '_'