:1234: Upgrade rancher to version 0.5.0
Also moved the rancher install code out to it's own file and added it
to .travis.yml as this only needs to run on Travis. The scripts are
linux specfic. The alternative would have been to detect the OS but dev
machines should have rancher as a prequisit if these scripts are being
run locally.